pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "fastparquet>=2025.12.0",
    "fire>=0.7.1",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pyarrow>=22.0.0",
    "pyqt5>=5.15.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "ruff>=0.11.9",
    "ty>=0.0.20",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "jupyterlab>=4.4.2",
    "ruff>=0.11.9",
]
